+++ package.json
@@ -19,10 +19,10 @@
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "typeorm": "typeorm-ts-node-commonjs",
-    "migration:generate": "pnpm ts-node -r tsconfig-paths/register node_modules/typeorm/cli.js migration:generate -d src/typeorm-cli.config.ts",
-    "migration:run": "pnpm ts-node -r tsconfig-paths/register node_modules/typeorm/cli.js migration:run -d typeorm-cli.ts",
-    "migration:revert": "pnpm ts-node -r tsconfig-paths/register node_modules/typeorm/cli.js migration:revert -d typeorm-cli.ts"
+    "typeorm": "ts-node -r tsconfig-paths/register --project tsconfig.json node_modules/typeorm/cli.js -d src/typeorm-cli.config.cjs",
+    "migrate": "ts-node -r tsconfig-paths/register --project tsconfig.json node_modules/typeorm/cli.js -d src/typeorm-cli.config.cjs migration:run",
+    "migrate:generate": "ts-node -r tsconfig-paths/register --project tsconfig.json node_modules/typeorm/cli.js -d src/typeorm-cli.config.cjs migration:generate",
+    "migrate:revert": "ts-node -r tsconfig-paths/register --project tsconfig.json node_modules/typeorm/cli.js -d src/typeorm-cli.config.cjs migration:revert"
   },
   "dependencies": {
     "@nestjs/common": "^11.0.1",
